# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant Ordering & Table Reservation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 1 */
  full_name: string;
  /** @nullable */
  phone?: string | null;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  customer: 'customer',
  staff: 'staff',
  kitchen_staff: 'kitchen_staff',
  manager: 'manager',
  owner: 'owner',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  email: string;
  full_name: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  created_at: string;
  is_active: boolean;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  customer: 'customer',
  staff: 'staff',
  kitchen_staff: 'kitchen_staff',
  manager: 'manager',
  owner: 'owner',
  admin: 'admin',
} as const;

export interface UserUpdate {
  full_name?: string;
  /** @nullable */
  phone?: string | null;
  is_active?: boolean;
  role?: UserUpdateRole;
}

export type MenuItemCategory = typeof MenuItemCategory[keyof typeof MenuItemCategory];


export const MenuItemCategory = {
  appetizer: 'appetizer',
  main: 'main',
  dessert: 'dessert',
  drink: 'drink',
  side: 'side',
  special: 'special',
} as const;

export interface MenuItem {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: MenuItemCategory;
  /** @nullable */
  image_url?: string | null;
  is_available: boolean;
  sort_order: number;
  created_at: string;
  updated_at: string;
}

export type MenuItemInputCategory = typeof MenuItemInputCategory[keyof typeof MenuItemInputCategory];


export const MenuItemInputCategory = {
  appetizer: 'appetizer',
  main: 'main',
  dessert: 'dessert',
  drink: 'drink',
  side: 'side',
  special: 'special',
} as const;

export interface MenuItemInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  price: number;
  category: MenuItemInputCategory;
  /** @nullable */
  image_url?: string | null;
  is_available?: boolean;
  sort_order?: number;
}

export type MenuItemUpdateCategory = typeof MenuItemUpdateCategory[keyof typeof MenuItemUpdateCategory];


export const MenuItemUpdateCategory = {
  appetizer: 'appetizer',
  main: 'main',
  dessert: 'dessert',
  drink: 'drink',
  side: 'side',
  special: 'special',
} as const;

export interface MenuItemUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  price?: number;
  category?: MenuItemUpdateCategory;
  /** @nullable */
  image_url?: string | null;
  is_available?: boolean;
  sort_order?: number;
}

export interface ActivityItem {
  id: string;
  description: string;
  timestamp: string;
  type: string;
}

export interface CustomerDashboard {
  role: string;
  total_orders: number;
  pending_reservations: number;
  loyalty_points: number;
  recent_activity: ActivityItem[];
}

export interface StaffDashboard {
  role: string;
  active_tables: number;
  pending_orders: number;
  todays_reservations: number;
  recent_activity: ActivityItem[];
}

export interface KitchenDashboard {
  role: string;
  orders_in_queue: number;
  orders_completed_today: number;
  avg_prep_time_minutes: number;
  recent_activity: ActivityItem[];
}

export interface ManagerDashboard {
  role: string;
  total_revenue_today: number;
  total_orders_today: number;
  staff_on_duty: number;
  pending_issues: number;
  recent_activity: ActivityItem[];
}

export interface OwnerDashboard {
  role: string;
  total_revenue_month: number;
  total_orders_month: number;
  total_customers: number;
  total_staff: number;
  recent_activity: ActivityItem[];
}

export interface AdminDashboard {
  role: string;
  total_users: number;
  active_sessions: number;
  system_status: string;
  recent_activity: ActivityItem[];
}

export type ListMenuItemsParams = {
category?: string;
available_only?: boolean;
};