# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant Ordering & Table Reservation API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemCategory } from './menuItemCategory';

export interface MenuItem {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  category: MenuItemCategory;
  /** @nullable */
  image_url?: string | null;
  is_available: boolean;
  sort_order: number;
  created_at: Date;
  updated_at: Date;
}