# lib/api-zod/src/generated/types/menuItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Restaurant Ordering & Table Reservation API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItemInputCategory } from './menuItemInputCategory';

export interface MenuItemInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  description?: string | null;
  /** @minimum 0 */
  price: number;
  category: MenuItemInputCategory;
  /** @nullable */
  image_url?: string | null;
  is_available?: boolean;
  sort_order?: number;
}